cp_library::binsearch
Function first_none

pub fn first_none<F, N, T>(l: N, r: N, predicate: F) -> (N, Option<T>)
where
    N: BinarySearchable,
    F: Fn(N) -> Option<T>,

Returns the first value in $[l, r)$ for which predicate returns None, or $r$ if none exists.

Furthermore, it returns predicate(x) for the greatest $x$ for which it is not None, or None if no such $x$ exists.

This is useful when an explicit construction is needed.

Conditions: there must exist a value $x_0$ such that predicate(x) return Some if and only if $x \ge x_0$.

Complexity: $\mathcal{O}(\log (r - l))$ calls to predicate.

Examples

use cp_library::binsearch::first_none;

let check = |val| if val * val < 6 { Some(val * val) } else { None };
let (fst, proof) = first_none(0, 15, check);

assert_eq!(fst, 3);
assert_eq!(proof, Some(4));

use cp_library::binsearch::first_none;

let check = |val| if val * val < 6 { Some(val * val) } else { None };
let (fst, proof) = first_none(5, 15, check);

assert_eq!(fst, 5);
assert_eq!(proof, None);