cp_library::inout
Struct InputReader

pub struct InputReader<R: Read> { /* private fields */ }

Buffered token-based input reader.

Provides a token-by-token input reader over any Read type, with automatic parsing into any FromStr type.

Implementations

impl<R: Read> InputReader<R>

pub fn new(reader: R) -> Self

Builds an input reader over a given reader, consuming it.

Examples

use std::io;
use cp_library::inout::InputReader;

let mut reader = InputReader::new(io::stdin());

pub fn get<T: FromStr>(&mut self) -> T
where <T as FromStr>::Err: Debug,

Extracts a single token and parses into a FromStr type.

Examples

use cp_library::inout::InputReader;

let mut reader = InputReader::new("123 abc".as_bytes());
assert_eq!(reader.get::<i32>(), 123);
assert_eq!(reader.get::<String>(), String::from("abc"));

Panics

If the next token fails to be parsed into T.

use cp_library::inout::InputReader;

let mut reader = InputReader::new("abc".as_bytes());
let x: i32 = reader.get();

pub fn get_vec<T: FromStr>(&mut self, size: usize) -> Vec<T>
where <T as FromStr>::Err: Debug,

Extracts size tokens of the same FromStr type and collect them in a Vec.

Examples

use cp_library::inout::InputReader;

let mut reader = InputReader::new("123 456 789".as_bytes());
assert_eq!(reader.get_vec::<i32>(3), vec![123, 456, 789]);

Panics

If any of the next size token fail to be parsed into T

use cp_library::inout::InputReader;

let mut reader = InputReader::new("123 456 abc".as_bytes());
let x = reader.get_vec::<i32>(3);