cp_library::math::algebra
Trait Abelian

pub trait Abelian: Group { }

An Abelian group must satisfy commutativity

This is marker trait, used to indicate that a Group is commutative

This trait is implemented for all signed builtin numeric types, with addition as the operation

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.