cp_library::math::algebra
Trait Monoid

pub trait Monoid: Semigroup {
    const ID: Self;
}

A Monoid must have a neutral element.

This trait is implemented for all builtin numeric types, with addition as the operation. Furthermore, if T is Semigroup, then Option<T> is automatically Monoid with None as the neutral element.

Required Associated Constants

const ID: Self

The neutral element of the monoid

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.