cp_library

Module string

String utilities

Modules

main_lorentz

Main-Lorentz algorithm for longest square subsequence

Structs

DBF

Cyclic dictionary basic factor of a string over an alphabet T.

Functions

lcp

Constructs the Suffix array and LCP array of a slice of T

suffix_array

Constructs the Suffix array of a slice of T

z_array

Calculate Z-function of a given slice of an Eq type T.