cp_library::string::main_lorentz

Function main_lorentz

pub fn main_lorentz<T: Eq>(arr: &[T]) -> Vec<MainLorentzRepetition>

Calculate all squared substrings (in compressed format) of a given slice of an Eq type T.

A squared string is a string which can be obtained by concatenating two identical strings.

Warning: This is not thoroughly tested since there are no (easily accessible) online judges with a task suitable for testing.

Complexity: $\mathcal{O}(N \log N)$ comparisons where:

$N$ is the size of the slice.